@@ -15,7 +15,7 @@
   ],
   "license": "BSD-3-Clause",
   "require": {
-    "php": ">=7.3.0",
+    "php": ">=7.4.0",
     "ext-mbstring": "*",
     "ext-json": "*"
   },
