@@ -16,7 +16,7 @@ class BoolMapper implements MapperInterface
     use SharedMapperTrait;
 
     /**
-     * Convert complex data (like object) to simpe data (scalar, null, array)
+     * Convert complex data (like object) to simple data (scalar, null, array)
      * @param $complex
      * @return bool
      * @throws SerializerException
